requirements.txt
google-adk
google-generativeai==0.8.5
psutil==5.9.5
python-dotenv==1.1.0
deprecated
pydantic